requirements.txt
numpy==1.20.3
Pillow==9.3.0
qrcode==7.3.1
scikit_image==0.19.3
